Fix calculation of 32-bit immediates via imm
The manual states that normally Type B instructions sign extend their
16-bit immediate to a 32 bit value, which is then used in computation.

It goes on to say that the IMM instruction can override this behavior
and provide the upper 16 bits of the 32-bit immediate for the next
instruction.

In the CTF binaries I looked at, this seems to mean the following
instruction should no longer sign extend its 16-bit immediate.